Completed first version of emtn.
Several basic unit tests are provided for initial confirmation that it
is functioning as expected.

One issue does still exist. Shutdown of emtn_prog seems to leave the
underlying leveldb (in ./priv/mtn/idx_store) in a locked or bad state.
This means that running `rebar eunit` twice will fail w/ a *timeout*.